Introduce integration tests fixture object
The fixture applies the operator yaml on `SetUp` and deletes it on
`TearDown`. Thus we could use the fixture in various context, for
example we could have a "create" suite that has multiple _parallel_ `It`s
that verify various aspects of the cfapi getting reconciled. The cfapi
resource should be applied just once in `SynchronizedBeforeSuite`

We could also have a "delete" suite that verifies the deletion
behaviour. Its `SynchronizedBeforeSuite` should apply the cfapi
resource, await its readiness and delete it. Parallel `It`s would verify
various aspects of the deletion behaviour

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>