WIP: Reimplement cfapi controller
The CFAPI controller implementation is quite complex and we believe it
is not worth refactoring it. Instead, we are going to reimplement it.

We have pulled the patching reconciler "framework" from Korifi and we
will base the new implementation on it.

We have also temporarily pinned `golang.org/x/tools` to `v0.37.0` to
workaround https://github.com/maxbrunsfeld/counterfeiter/issues/344

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>