Spike: use the docker registry internal secret
This spike proves that the CFAPI operator could use the internal kyma docker registry
secret when configuring Korifi. Thus we can lift off the requirement
that the docker registry must have its exnternal access enabled.

The prepare-kind script still enables the external access for the
registry but that is purely needed in order to be able to push images
the script itself builds. cf pushing apps works perfectly fine when
Korifi is configured with the internal secret.

All tha makes sense as the kpack builder and the Korifi API both live on
the same cluster, i.e. external access is not needed.