Use `pushRegAddr` in kyma docker registry secret as registy url
On a real kyma, the registry secret contains two entries in the `auths`
map: one for the pull url, and one for the push. Provided that korifi
only pushes to the registry, we ensure that we pick push url to
configure korifi api.

For custom registry secrets we just pick the first entry on the `auths`
map

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>