Use the docker registry extrernal secret for registry config
The internal URL (`*.svc.cluster.local`) cannot be used for workload pod
images on a real Kyma cluster as those images are loaded by the kubelet.
The kubelet itself runs on the node in its network namespace, therefore
it cannot resolve `svc.cluster.local`

When deploying on kind (dev environment or integration tests), we
unfortunately cannot use the external secret as the external url is not
accessible from within the korifi api/controllers pods. Therefore for
kind deployments we configure the cfapi resource to use the internal
docker registry secret AND disable the secret propagation (the kyma
docker registry does it for us)