Option to enable tmate so ssh into actions container
The `action-tmate` action sets up a detached tmate session if the
`debug` workflow option is enabled. Enabling this option is possible
when the workflow is run manually.

The output of the action outputs an `ssh` command one can use to attach
to the actions container. in order to make debugging easier, we also
install k9s